rename alert types to rule types
